# sacdia-api-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/users/{userId}/emergency-contacts": {
      "post": {
        "operationId": "EmergencyContactsController_create_v1",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmergencyContactDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contacto creado"
          },
          "400": {
            "description": "Máximo de contactos alcanzado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear contacto de emergencia (máximo 5)",
        "tags": [
          "emergency-contacts"
        ]
      },
      "get": {
        "operationId": "EmergencyContactsController_findAll_v1",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de contactos"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar contactos de emergencia del usuario",
        "tags": [
          "emergency-contacts"
        ]
      }
    },
    "/v1/users/{userId}/emergency-contacts/{contactId}": {
      "get": {
        "operationId": "EmergencyContactsController_findOne_v1",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contactId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contacto encontrado"
          },
          "404": {
            "description": "Contacto no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener un contacto específico",
        "tags": [
          "emergency-contacts"
        ]
      },
      "patch": {
        "operationId": "EmergencyContactsController_update_v1",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contactId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmergencyContactDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contacto actualizado"
          },
          "404": {
            "description": "Contacto no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar contacto de emergencia",
        "tags": [
          "emergency-contacts"
        ]
      },
      "delete": {
        "operationId": "EmergencyContactsController_remove_v1",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contactId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contacto eliminado"
          },
          "404": {
            "description": "Contacto no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar contacto de emergencia (soft delete)",
        "tags": [
          "emergency-contacts"
        ]
      }
    },
...
    "/v1/finances/categories": {
      "get": {
        "description": "Lista todas las categorías de ingresos y egresos",
        "operationId": "FinancesController_getCategories_v1",
        "parameters": [
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "0=Ingresos, 1=Egresos",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de categorías"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar categorías financieras",
        "tags": [
          "finances"
        ]
      }
    },
    "/v1/clubs/{clubId}/finances": {
      "get": {
        "description": "Obtiene todos los movimientos de las instancias del club",
        "operationId": "FinancesController_findByClub_v1",
        "parameters": [
          {
            "name": "clubId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "month",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "clubTypeId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de movimientos"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar movimientos financieros del club",
        "tags": [
          "finances"
        ]
      },
      "post": {
        "description": "Crea un nuevo ingreso o egreso (requiere rol de tesorería)",
        "operationId": "FinancesController_create_v1",
        "parameters": [
          {
            "name": "clubId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFinanceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Movimiento creado"
          },
          "403": {
            "description": "Permisos insuficientes"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear movimiento financiero",
        "tags": [
          "finances"
        ]
      }
    },
    "/v1/clubs/{clubId}/finances/summary": {
      "get": {
        "description": "Obtiene el resumen de ingresos, egresos y balance",
        "operationId": "FinancesController_getSummary_v1",
        "parameters": [
          {
            "name": "clubId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "month",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resumen financiero"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Resumen financiero del club",
        "tags": [
          "finances"
        ]
      }
    },
    "/v1/finances/{financeId}": {
      "get": {
        "operationId": "FinancesController_findOne_v1",
        "parameters": [
          {
            "name": "financeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Movimiento encontrado"
          },
          "404": {
            "description": "Movimiento no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener movimiento por ID",
        "tags": [
          "finances"
        ]
      },
      "patch": {
        "operationId": "FinancesController_update_v1",
        "parameters": [
          {
            "name": "financeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFinanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Movimiento actualizado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar movimiento",
        "tags": [
          "finances"
        ]
      },
      "delete": {
        "operationId": "FinancesController_remove_v1",
        "parameters": [
          {
            "name": "financeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Movimiento desactivado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Desactivar movimiento",
        "tags": [
          "finances"
        ]
      }
    }
  },
  "info": {
    "title": "SACDIA API",
    "description": "## Sistema de Administración de Clubes de Conquistadores y Aventureros\n\n### Módulos Disponibles\n- **Auth**: Autenticación con Supabase + JWT\n- **Users**: Gestión de perfiles de usuario\n- **Catalogs**: Catálogos de referencia\n- **Clubs**: Gestión de clubes e instancias\n- **Classes**: Clases progresivas\n- **Honors**: Especialidades\n- **Activities**: Actividades de club\n- **Finances**: Control financiero",
    "version": "2.2.0",
    "contact": {
      "name": "SACDIA Team",
      "url": "https://sacdia.app",
      "email": "dev@sacdia.app"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateEmergencyContactDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "example": "María García López"
          },
          "relationship_type": {
            "type": "number",
            "example": 1,
            "description": "ID del tipo de relación (actualmente Int, pendiente migración a UUID)"
          },
          "phone": {
            "type": "string",
            "maxLength": 20,
            "example": "+52 55 1234 5678"
          },
          "primary": {
            "type": "boolean",
            "example": true,
            "description": "Si es el contacto principal"
          }
        },
        "required": [
          "name",
          "relationship_type",
          "phone"
        ]
      },
      "UpdateEmergencyContactDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "example": "María García López"
          },
          "relationship_type": {
            "type": "number",
            "example": 1,
            "description": "ID del tipo de relación (actualmente Int, pendiente migración a UUID)"
          },
          "phone": {
            "type": "string",
            "maxLength": 20,
            "example": "+52 55 1234 5678"
          },
          "primary": {
            "type": "boolean",
            "example": true,
            "description": "Si es el contacto principal"
          }
        }
      },
...
      "CreateFinanceDto": {
        "type": "object",
        "properties": {
          "year": {
            "type": "number",
            "description": "Año del movimiento"
          },
          "month": {
            "type": "number",
            "minimum": 1,
            "description": "Mes del movimiento (1-12)"
          },
          "amount": {
            "type": "number",
            "description": "Monto del movimiento (en centavos)"
          },
          "description": {
            "type": "string",
            "description": "Descripción del movimiento"
          },
          "club_type_id": {
            "type": "number",
            "description": "Tipo de club (1=Aventureros, 2=Conquistadores, 3=GM)"
          },
          "finance_category_id": {
            "type": "number",
            "description": "ID de la categoría financiera"
          },
          "finance_date": {
            "type": "string",
            "description": "Fecha del movimiento"
          },
          "club_adv_id": {
            "type": "number",
            "description": "ID de la instancia de Aventureros"
          },
          "club_pathf_id": {
            "type": "number",
            "description": "ID de la instancia de Conquistadores"
          },
          "club_mg_id": {
            "type": "number",
            "description": "ID de la instancia de Guías Mayores"
          }
        },
        "required": [
          "year",
          "month",
          "amount",
          "club_type_id",
          "finance_category_id",
          "finance_date"
        ]
      },
      "UpdateFinanceDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "finance_category_id": {
            "type": "number"
          },
          "finance_date": {
            "type": "string"
          }
        }
      }
    }
  }
}